Update FishSimulator.java
I started working on the fishing methods that determine whether or not a fish was found and if the number of fishing attempts was smaller than the max amount. I also tried to create a catch fish method that takes user input but I am not sure how to implement it within the object class and carry over to the runner class.